scoreカラムの型をfloatに修正

+++ db/schema.rb
@@ -107,7 +107,7 @@ ActiveRecord::Schema.define(version: 2021_05_12_022701) do
 
   create_table "reviews", charset: "utf8", force: :cascade do |t|
     t.string "comment"
-    t.integer "score", null: false
+    t.float "score", null: false
     t.bigint "consignment_side_user_id"
     t.bigint "post_id"
     t.datetime "created_at", precision: 6, null: false
